LibC: Fix compiler warnings
It can now compile with almost all compiler warnings turned on without
producing any warnings.

+++ userland/libc/include/tb/sb.h
@@ -16,7 +16,7 @@ void sb_free(struct sb *ctx);
 void sb_reset(struct sb *ctx);
 int sb_isempty(const struct sb *ctx);
 void sb_append_char(struct sb *ctx, char c);
-int sb_delete_right(struct sb *ctx, int n);
+int sb_delete_right(struct sb *ctx, size_t n);
 void sb_append(struct sb *ctx, const char *s);
 void sb_append_sv(struct sb *ctx, struct sv sv);
 void sb_prepend_sv(struct sb *ctx, struct sv sv);
